feat: implement release build job and refactor build job creation
