+++ crates/forge_ci/src/jobs/release_build_job.rs
@@ -1,0 +1,45 @@
+use gh_workflow_tailcall::*;
+
+use crate::{jobs::apt_get_install, release_matrix::ReleaseMatrix};
+
+pub fn release_build_job() -> Job {
+    Job::new("build-release")
+        .strategy(Strategy {
+            fail_fast: None,
+            max_parallel: None,
+            matrix: Some(ReleaseMatrix::default().into()),
+        })
+        .runs_on("${{ matrix.os }}")
+        .permissions(
+            Permissions::default()
+                .contents(Level::Write)
+                .pull_requests(Level::Write),
+        )
+        .add_step(Step::uses("actions", "checkout", "v4"))
+        // Install Rust with cross-compilation target
+        .add_step(
+            Step::uses("taiki-e", "setup-cross-toolchain-action", "v1")
+                .with(("target", "${{ matrix.target }}")),
+        )
+        // Explicitly add the target to ensure it's available
+        .add_step(Step::run("rustup target add ${{ matrix.target }}").name("Add Rust target"))
+        // Build add link flags
+        .add_step(
+            Step::run(r#"echo "RUSTFLAGS=-C target-feature=+crt-static" >> $GITHUB_ENV"#)
+                .if_condition(Expression::new(
+                    "!contains(matrix.target, '-unknown-linux-gnu')",
+                )),
+        )
+        .add_step(
+            Step::run(apt_get_install(&[
+                "gcc-aarch64-linux-gnu",
+                "musl-tools",
+                "musl-dev",
+                "pkg-config",
+                "libssl-dev",
+            ]))
+            .if_condition(Expression::new(
+                "contains(matrix.target, '-unknown-linux-musl')",
+            )),
+        )
+}
